Preserve row order, convert NoData to NA
In base R, merge does not preserve the row order of the 'left' dataframe
in a left join (unlike in Pandas). Adding an 'id' column, sorting by it
after the merge, and removing it later solves this.

Also, the lookup tables contain "NoData" for empty cells which need to
be converted to NA for the user.